From TextView to RecyclerView
Changed the display of SQLite database data from TextView to RecyclerView.

Added an SQLiteAdapter and a RecyclerView item layout.

Changed the TextView element to a RecyclerView element in activity_main.xml.

Changed the viewData logic to support RecyclerViews instead of TextViews, adding a new RecyclerView field for this purpose.

In myDBAdapter.java, changed getData() return type from String to String[]. Also made the stringBuffer variable an ArrayList<String> to support dynamically adding new strings instead of the previous method of appending a substring to a single String variable for every piece of data there was.